Replace datapersister with processors
