+++ backend/packages/Core/config/services.yaml
@@ -27,8 +27,9 @@ services:
             - '@App\Core\OpenApi\OpenApiFactory.inner'
         autoconfigure: false
 
-    App\Core\DataPersister\UserDataPersister:
-        decorates: api_platform.doctrine.orm.data_persister
+    App\Core\State\UserProcessor:
+        bind:
+            $decorated: '@api_platform.doctrine.orm.state.persist_processor'
 
     App\Core\Service\MailerService:
         bind:
