--- backend/packages/Fishbowl/src/DataPersister/FishbowlDataPersister.php
@@ -1,65 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Stooa codebase.
- *
- * (c) 2020 - present Runroom SL
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace App\Fishbowl\DataPersister;
-
-use ApiPlatform\Core\DataPersister\ContextAwareDataPersisterInterface;
-use App\Core\Service\MailerService;
-use App\Fishbowl\Entity\Fishbowl;
-
-class FishbowlDataPersister implements ContextAwareDataPersisterInterface
-{
-    public function __construct(
-        private readonly ContextAwareDataPersisterInterface $decorated,
-        private readonly MailerService $mailerService
-    ) {
-    }
-
-    /**
-     * @param mixed $data
-     * @param mixed[] $context
-     */
-    public function supports($data, array $context = []): bool
-    {
-        return $this->decorated->supports($data, $context);
-    }
-
-    /**
-     * @param mixed $data
-     * @param mixed[] $context
-     *
-     * @return object|void
-     */
-    public function persist($data, array $context = [])
-    {
-        $result = $this->decorated->persist($data, $context);
-
-        if ($data instanceof Fishbowl && (
-            ($context['collection_operation_name'] ?? null) === 'post' ||
-            ($context['graphql_operation_name'] ?? null) === 'create')
-        ) {
-            $this->mailerService->sendFishbowlCreatedEmail($data);
-        }
-
-        return $result;
-    }
-
-    /**
-     * @param mixed $data
-     * @param mixed[] $context
-     */
-    public function remove($data, array $context = []): void
-    {
-        $this->decorated->remove($data, $context);
-    }
-}
